+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## Version 0.03.2
+
+* Add sphinx documentation
+* Add Travis CI support
+* Add `list_tables` method to `Table` class
+* Make column name/primary key/autoincrement parsing more robust
+* Check if SQLite table exists on `Table` initialization
+* Allow tests to be run from within `tests` directory
+
+## Version 0.03.1
+
+* Fix `pip install` installation bug
+
 ## Version 0.03
 
 * Fix an issue in `Table` initialization from `DataFrame` (#5)
@@ -8,10 +21,6 @@
 * Create testing metaclass to modify test case docstrings to include
   class name
 * Add `Makefile` and other minor installation modifications
-
-# Version 0.03.1
-
-* Fix `pip install` installation bug
 
 ## Version 0.02